feat: change echo commando for send-message
